# [Go] Optimise HLL union merge (HLLtoHLL)