```
goos: darwin
goarch: arm64
pkg: github.com/apache/datasketches-go/hll
                  │   old.txt    │              new-2.txt              │
                  │    sec/op    │   sec/op     vs base                │
HLLMerge/merge-10   1322.5n ± 1%   676.4n ± 0%  -48.85% (p=0.000 n=10)
```